Split cli.py into multiple smaller files
